+++ Assets/Scripts/Product.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 public class Product : ScriptableObject
 {
     public string label;
-    public int consumptionSpeed;
-    public int initialValue;
+    public float consumptionSpeed;
+    public float initialValue;
 }
